refactor: rename BaseResponse -> Response, Response->SimpleResponse
I think it is a lot more intuitive if llm.call always returns `llm.Response`, which has everything you expect defined there, and what users mostly think about. Then in the specific case where they are using tools, they can specify whether they expect `llm.SimpleResponse` or `llm.ContextResponse`. This is more intuitive than making `llm.BaseResponse` the class they are always expecting (and more accurate than treating the SimpleResponse as the default, because that isn't consistently accurate).
