+++ python/mirascope/llm/__init__.py
@@ -44,7 +44,9 @@ from .response_formatting.decorator import response_format
 from .responses import (
     AsyncStream,
     AsyncStructuredStream,
+    ContextResponse,
     Response,
+    SimpleResponse,
     Stream,
     StreamChunk,
     StructuredStream,
@@ -66,6 +68,7 @@ __all__ = [
     "Client",
     "Content",
     "Context",
+    "ContextResponse",
     "Document",
     "Image",
     "Message",
@@ -73,6 +76,7 @@ __all__ = [
     "Prompt",
     "Response",
     "ResponseFormat",
+    "SimpleResponse",
     "Stream",
     "StreamChunk",
     "StructuredAgent",
