+++ python/mirascope/llm/responses/__init__.py
@@ -4,7 +4,6 @@ from .async_context_stream import AsyncContextStream
 from .async_context_structured_stream import AsyncContextStructuredStream
 from .async_stream import AsyncStream
 from .async_structured_stream import AsyncStructuredStream
-from .base_response import BaseResponse
 from .content import ContextResponseContent, ResponseContent
 from .context_response import ContextResponse
 from .context_stream import ContextStream
@@ -12,6 +11,7 @@ from .context_stream_chunk import ContextStreamChunk
 from .context_structured_stream import ContextStructuredStream
 from .finish_reason import FinishReason
 from .response import Response
+from .simple_response import SimpleResponse
 from .stream import Stream
 from .stream_chunk import StreamChunk
 from .structured_stream import StructuredStream
@@ -22,15 +22,15 @@ __all__ = [
     "AsyncContextStructuredStream",
     "AsyncStream",
     "AsyncStructuredStream",
-    "BaseResponse",
     "ContextResponse",
     "ContextResponseContent",
     "ContextStream",
     "ContextStreamChunk",
     "ContextStructuredStream",
     "FinishReason",
-    "Response",
     "ResponseContent",
+    "SimpleResponse",
+    "SimpleResponse",
     "Stream",
     "StreamChunk",
     "StructuredStream",
