+++ python/mirascope/llm/responses/context_response.py
@@ -13,15 +13,15 @@ from typing_extensions import TypeVar
 
 from ..tools import ContextTool
 from ..types import Jsonable
-from .base_response import BaseResponse
 from .content import ResponseContent
+from .response import Response
 
 DepsT = TypeVar("DepsT", default=None)
 T = TypeVar("T", bound=object | None, default=None)
 
 
 @dataclass
-class ContextResponse(BaseResponse[ResponseContent, T], Generic[DepsT, T]):
+class ContextResponse(Response[ResponseContent, T], Generic[DepsT, T]):
     """The response generated by an LLM."""
 
     tools: Sequence[ContextTool[..., Jsonable, DepsT]]
