+++ python/mirascope/llm/responses/simple_response.py
@@ -1,0 +1,30 @@
+"""Interfaces for LLM responses.
+
+This module defines interfaces for the responses returned by language models,
+including methods for formatting the response according to a specified format.
+"""
+
+from collections.abc import Sequence
+from dataclasses import dataclass
+
+from typing_extensions import TypeVar
+
+from ..tools import Tool
+from ..types import Jsonable
+from .content import ResponseContent
+from .response import Response
+
+T = TypeVar("T", bound=object | None, default=None)
+
+
+@dataclass
+class SimpleResponse(Response[ResponseContent, T]):
+    """The response generated by an LLM."""
+
+    tools: Sequence[Tool[..., Jsonable]]
+    """The tools the LLM wants called on its behalf, if any."""
+
+    @property
+    def tool(self) -> Tool[..., Jsonable] | None:
+        """Returns the first tool used in the response, if any."""
+        raise NotImplementedError()
